# PERF: df.groupby(categorical)