@topper-123 2022-11-09
Wow, very nice!

One question, that's only semi-related to this: why do we even use `reorder_categories` in this groupby? it does seem unrelated to especially as we're only using one dtype (if we were combining two dtypes, I would understand it).

Could we get an additional (minor) perf. improvement by avoiding `reorder_categories` in groupby in cases like this?