@lukemanley 2022-11-09
I'm going to guess the rationale for matching the original order of the categories is to keep the result categories dtype equivalent to the input categories dtype. I think we could add a short-circuit that checks to see if they happen to be equal already and if so skip `reorder_categories`. However, I just tried that locally and could not measure any noticeable improvement in performance.