@topper-123 2022-11-09
Ok. Thanks for trying it out. This performance improvement is already really great as is :-)